deduplicate types
