+++ packages/payload/src/admin/RichText.ts
@@ -14,16 +14,15 @@ import type {
   Validate,
 } from '../fields/config/types.js'
 import type { SanitizedGlobalConfig } from '../globals/config/types.js'
-import type { RequestContext, TypedFallbackLocale } from '../index.js'
+import type { RequestContext, TypedFallbackLocale, TypedUser } from '../index.js'
 import type { JsonObject, PayloadRequest, PopulateType } from '../types/index.js'
 import type { RichTextFieldClientProps, RichTextFieldServerProps } from './fields/RichText.js'
-import type { FormStateWithoutComponents } from './forms/Form.js'
 import type {
-  ClientFieldSchemaMap,
-  FieldDiffClientProps,
-  FieldDiffServerProps,
-  FieldSchemaMap,
-} from './types.js'
+  AddFieldStatePromiseArgs,
+  CalculateDefaultValuesIterateFieldsArgs,
+  FormStateIterateFieldsArgs,
+} from './forms/Form.js'
+import type { FieldDiffClientProps, FieldDiffServerProps, FieldSchemaMap } from './types.js'
 
 export type AfterReadRichTextHookArgs<
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
@@ -220,83 +219,11 @@ type RichTextAdapterBase<
    * Nested field paths use the node ID as a bridge: `{path}.{nodeId}.{fieldName}`.
    * The node ID flattens the freeform tree — the form state doesn't care about tree depth or position.
    */
-  buildFormState?: (args: {
-    /**
-     * The rich text field's data value. May be the original data (initial load) or include
-     * current nested field values merged in by `reduceFieldsToValues` + unflatten (on change).
-     */
-    data: any
-    fieldSchemaMap: FieldSchemaMap
-    /**
-     * The `iterateFields` function from `@payloadcms/ui`. Call this for each node's fields
-     * with the appropriate `parentPath` and `parentSchemaPath`.
-     */
-    iterateFields: (args: {
-      addErrorPathToParent: (fieldPath: string) => void
-      anyParentLocalized?: boolean
-      blockData: any
-      clientFieldSchemaMap?: ClientFieldSchemaMap
-      collectionSlug?: string
-      data: any
-      fields: any[]
-      fieldSchemaMap: FieldSchemaMap
-      filter?: any
-      forceFullValue?: boolean
-      fullData: any
-      id?: number | string
-      includeSchema?: boolean
-      mockRSCs?: any
-      omitParents?: boolean
-      operation: 'create' | 'update'
-      parentIndexPath: string
-      parentPassesCondition?: boolean
-      parentPath: string
-      parentSchemaPath: string
-      permissions: any
-      preferences?: any
-      previousFormState: any
-      readOnly?: boolean
-      renderAllFields: boolean
-      renderFieldFn: any
-      req: PayloadRequest
-      select?: any
-      selectMode?: any
-      skipConditionChecks?: boolean
-      skipValidation?: boolean
-      state?: FormStateWithoutComponents
-    }) => Promise<void>
-    /**
-     * Common args to forward to `iterateFields`. These come from the parent `addFieldStatePromise` context.
-     */
-    iterateFieldsArgs: {
-      addErrorPathToParent: (fieldPath: string) => void
-      anyParentLocalized?: boolean
-      clientFieldSchemaMap?: ClientFieldSchemaMap
-      collectionSlug?: string
-      fieldSchemaMap: FieldSchemaMap
-      filter?: any
-      forceFullValue?: boolean
-      fullData: any
-      id?: number | string
-      includeSchema?: boolean
-      mockRSCs?: any
-      omitParents?: boolean
-      operation: 'create' | 'update'
-      preferences?: any
-      previousFormState: any
-      readOnly?: boolean
-      renderAllFields: boolean
-      renderFieldFn: any
-      req: PayloadRequest
-      select?: any
-      selectMode?: any
-      skipConditionChecks?: boolean
-      skipValidation?: boolean
-      state: FormStateWithoutComponents
-    }
-    path: string
-    schemaPath: string
-  }) => Promise<void>
+  buildFormState?: (
+    args: {
+      iterateFields: (args: FormStateIterateFieldsArgs) => Promise<void>
+    } & AddFieldStatePromiseArgs<RichTextField>,
+  ) => Promise<void>
   /**
    * Apply default values to nested fields within the rich text value (e.g. blocks, inline blocks).
    *
@@ -308,18 +235,10 @@ type RichTextAdapterBase<
   calculateDefaultValues?: (args: {
     data: any
     id?: number | string
-    iterateFields: (args: {
-      data: any
-      fields: (Field | TabAsField)[]
-      id?: number | string
-      locale: string | undefined
-      req: PayloadRequest
-      siblingData: Record<string, unknown>
-      user: PayloadRequest['user']
-    }) => Promise<void>
+    iterateFields: (args: CalculateDefaultValuesIterateFieldsArgs<any>) => Promise<void>
     locale: string | undefined
     req: PayloadRequest
-    user: PayloadRequest['user']
+    user: TypedUser
   }) => Promise<void>
   /**
    * Provide a function that can be used to add items to the import map. This is useful for
