+++ packages/ui/src/forms/fieldSchemasToFormState/iterateFields.ts
@@ -1,85 +1,9 @@
-import type {
-  BuildFormStateArgs,
-  ClientFieldSchemaMap,
-  Data,
-  DocumentPreferences,
-  Field,
-  FieldSchemaMap,
-  FormState,
-  FormStateWithoutComponents,
-  PayloadRequest,
-  SanitizedFieldsPermissions,
-  SelectMode,
-  SelectType,
-  TabAsField,
-} from 'payload'
+import type { FormStateIterateFieldsArgs } from 'payload/internal'
 
 import { stripUnselectedFields } from 'payload'
 import { getFieldPaths } from 'payload/shared'
 
-import type { AddFieldStatePromiseArgs } from './addFieldStatePromise.js'
-import type { RenderFieldMethod } from './types.js'
-
 import { addFieldStatePromise } from './addFieldStatePromise.js'
-
-type Args = {
-  addErrorPathToParent: (fieldPath: string) => void
-  /**
-   * if any parents is localized, then the field is localized. @default false
-   */
-  anyParentLocalized?: boolean
-  /**
-   * Data of the nearest parent block, or undefined
-   */
-  blockData: Data | undefined
-  clientFieldSchemaMap?: ClientFieldSchemaMap
-  collectionSlug?: string
-  data: Data
-  fields: (Field | TabAsField)[]
-  fieldSchemaMap: FieldSchemaMap
-  filter?: (args: AddFieldStatePromiseArgs) => boolean
-  /**
-   * Force the value of fields like arrays or blocks to be the full value instead of the length @default false
-   */
-  forceFullValue?: boolean
-  fullData: Data
-  id?: number | string
-  /**
-   * Whether the field schema should be included in the state. @default false
-   */
-  includeSchema?: boolean
-  mockRSCs?: BuildFormStateArgs['mockRSCs']
-  /**
-   * Whether to omit parent fields in the state. @default false
-   */
-  omitParents?: boolean
-  /**
-   * operation is only needed for validation
-   */
-  operation: 'create' | 'update'
-  parentIndexPath: string
-  parentPassesCondition?: boolean
-  parentPath: string
-  parentSchemaPath: string
-  permissions: SanitizedFieldsPermissions
-  preferences?: DocumentPreferences
-  previousFormState: FormState
-  readOnly?: boolean
-  renderAllFields: boolean
-  renderFieldFn: RenderFieldMethod
-  req: PayloadRequest
-  select?: SelectType
-  selectMode?: SelectMode
-  /**
-   * Whether to skip checking the field's condition. @default false
-   */
-  skipConditionChecks?: boolean
-  /**
-   * Whether to skip validating the field. @default false
-   */
-  skipValidation?: boolean
-  state?: FormStateWithoutComponents
-}
 
 /**
  * Flattens the fields schema and fields data
@@ -117,7 +41,7 @@ export const iterateFields = async ({
   skipConditionChecks = false,
   skipValidation = false,
   state = {},
-}: Args): Promise<void> => {
+}: FormStateIterateFieldsArgs): Promise<void> => {
   const promises = []
 
   fields.forEach((field, fieldIndex) => {
